Move registry files from platform to src
